Code reorganisation to separate tskit from msprime.
Creates a new header 'trees.h' which contains all the non simulator
stuff from msprime.
